Ajout d'un bouléen pour savoir si c'est runnable la période peut maintenant être demandé par l'utilisateur

+++ testCapteur/src/nf33/realtime/apirtdroid/RTRunnable.java
@@ -11,6 +11,6 @@ public interface RTRunnable
 {
 	
 	public abstract void periodicEvent(long timeSinceLast);
-	public abstract void endConfiguration(Boolean isRunable, float frequence);
+	public abstract void endConfiguration(Boolean isRunable, Long frequence);
 
 }
